Validate 2012/2014 version/builds due known issue
Validate if server version is:
- sql 2012 and if have SP3 CU3 (Build 6537) or higher
- sql 2014 and if have SP2 (Build 5000) or higher
If the major version is the same but the build is lower, throws the
message:
"This SQL version has a known issue. Rebuilding an index clears any
existing row entry from sys.dm_db_index_usage_stats for that index.
Please refer to connect item:
https://connect.microsoft.com/SQLServer/feedback/details/739566/rebuilding-an-index-clears-stats-from-sys-dm-db-index-usage-stats"