Update of Copy-SqlJob to ignore maintenance plans by default
Found after performing work in client environment with this code that issues can occur if the jobs associated with maintenance plans are copied over and then later the maintenance plans are copied over.

It is best that the Copy-SqlJob only copy the jobs associated with the maintenance plans, if the maintenance plans are to be copied over at that time. The adjustment simply changes the "IgnoreMaintenancePlan" to "IncludeMaintenancePlan", also added help parameter/example.