Fix log was growing to the target size at once and not in chuncks
- The Alter() method was outside of WHILE cycle, this leads to the
ALTER() method only runs after which means that only runs ALTER() with
the full 'TotalTargetSizeMB'.  This problem was introduced on a earlier
commit.
- Fix the search for Lun available space. There was missing the final
backslash which leads to a wrong drive/lun validation. This is because
Split-Path removes the final backslash but the Get-DbaDiskSpace command
has it. If the file was on a LUN this will validate based on
RouteMountPoint/Drive and not based on LUN itself..
- Correct variable name for messagebox from '$SuggestLogIncrentSize' to
'$SuggestLogIncrementSize'
- Cleaned up some old comments.