Fixed Arithmetic Overflows
Some of my SQL Servers were causing arithmetic overflows grabbing the
space information.  The issue was in the SQL query itself.  I tried
changing Decimal(15,2) all the way up to (38,2) but it would not correct
the issue.  I chagned the decimals to floats.  Also there was a case
statement missing the converts that I had to add in UnusableSpaceMD.
This seems to work now on my servers.