Added Try block to server connect & fixed bug with NULL Full Text Indexe Sizes
The server connect code was not in a try block.  Script would crash out
if it couldn't find the server.  Also, sometimes Full Text indexes would
return NULL values.  The previous code was reusing values that were not
correct.  I had to check for NULL then set to zero.  The introduction of
a Round function on a previous update also threw errors with NULL
values.