New module - Find slowest stored procedure and sql statement executions per database.
Return the slowest executions of stored procedures as a whole and
individual SQL statements.  SQL Statements can be ad hoc or within the
stored procedures.  Default values:  100 results, with 100 executions or
more with a minimum execution time of 500ms.  These defaults are to
prevent massive amounts of data coming back in databases with thousands
or more query plans.