Update Issues template with examples
Making it a bit more fun and adding text that will hopefully catch folks eye so they will try to add more information to their issue submissions. Also added information around the logging system and how to find the logs to attach.

If it is a bit overboard we can scrap it.