Test-DbaSpn, local account fixes
Fixes #1366 (or, it *should*)

Changes:
- when the serviceaccount is detected as LocalSystem or is of type "NT
SERVICE\xxx" substitutes the object where to register the SPN with the
computer object.

Clusters should be fine, 'cause - theoretically - they should not run on
a local account.

Remains to be seen how the cmdlet handles instances with a MSAs, but I
don't have any available at hand to verify.