Make the functions work without job values
New-DbaAgentJob was cleaned in the parameter section. Two parameters are
added (schedule and scheduleid) to make it possible to attach a shared
schedule when the new job gets created. The return of the object was
also improved.

New-DbaAgentSchedule got a cleanup in the parameter section. It not
longer needs a job to create a schedule. The job parameters can still be
used to assign the schedule to a job.

Remove-DbaAgentJob cleanup in the parameter section and the job
parameter was set to be an object type instead of a string.

Remove-DbaAgentSchedule had the help improved and a cleanup in the
parameter section. It not longer relies on the job and has had a full
rewrite to support that.

Set-DbaAgentJob had a cleanup in the parameter section.  Two parameters
are added (schedule and scheduleid) to make it possible to attach a
shared schedule to a job.