Adjusted ComputerName to be a string
Adjusted ComputerName to be a string instead of the DbaInstanceParameter type. If you pass in a computer that does not exist it will throw an error that it cannot convert the value and then can't parse the instance name. Then output the expected warning messages.

Also adjusted the alias to match the example, SqlServer to SqlInstance.