Fix $Query variable check issue
Fix check issue with $Query parameter, as this variable is re-used in the event that multiple tables are copied the current check does not properly validate this scenario.  The $Query variable is written to in the first iteration of the loop and any subsequent iteration will result in the copy failing to run as expected.  Also corrected an issue with the truncation which resulted from the "Query" method being used which does not seem to exist, changed to "ExecuteNonQuery" method.