Remove Distribution from list of excluded DBs
Line 122 lists system databases which are to be excluded from process. I have removed distribution as it is perfectly valid to set this to any recovery model, replication is not affected by this. Also, distribution is a valid name for a user database, it is not reserved for replication.

I believe this could also be done for SSIS but I wasn't sure the reason this was listed so have left it for now.