Measure-DbaDiskSpaceRequirement strike 2
In response to wsmelton
I can't use Get-DbaDatabaseFile because it doesn't work with a French version of SQL Server AND It assume that files are on MountPoint of 1 character in length which is wrong.

For the progress bar, I will simply removed it for now.
I moved the private functions inside the begin block.
I will renamed the variables for CamelCasing.