updating gitignore file for docker-compose
Adding ignore for the docker-compose yaml file. In situations where folks want to use docker containers for testing the module and learning. Docker integration with VS Code and other tools, including docker, require the yaml file to exist within the project. (Future post to be written on this as well.)