see description
There is no reason to validate if a user has a login and what login type it is because if your user count is greater than 0 you hit your else statement and this validation in your filter is what excludes the user properly: ($User -contains $_.Name).

A user in a database that is authenticated with a certificate will never have a login so it will never match the name properly.