Update Remove-DbaOrphanUser.ps1 fix
Updating Remove-Orphan logic to determine users who are orphaned to be -eq 0 for the top, then else is that users are present. This functionality now matches Repair-Orphan, which was properly determining the orphans. Basically the if condition was inverted.
Uploaded due to GitHub desktop and VSCode failing authentication to push the change