Fixing Remove-DbaOrphanUser (#4767)
* Update Copy-DbaDbQueryStoreOption.ps1

Fixing parameter names to match what they are in $SourceQSConfig which came from Get-DbaDbQueryStoreOption, and splatting the parameters to get rid of the back-tick/grave accent syntax.

* Update Copy-DbaDbQueryStoreOption.ps1

Applied Invoke-DbatoolsFormatter to comply with OTSB formatting request

* Update Remove-DbaOrphanUser.ps1

Updating Remove-Orphan logic to determine users who are orphaned to be -eq 0 for the top, then else is that users are present.  This functionality now matches Repair-Orphan, which was properly determining the orphans.  Basically the if condition was inverted.

* Update Remove-DbaOrphanUser.ps1 fix

Updating Remove-Orphan logic to determine users who are orphaned to be -eq 0 for the top, then else is that users are present. This functionality now matches Repair-Orphan, which was properly determining the orphans. Basically the if condition was inverted.
Uploaded due to GitHub desktop and VSCode failing authentication to push the change

* Accidentially checked in here

Didn't want the file here, trying to update in functions, but failed due to connectivity issues.

* dropping improperly placed file