Get-DbaDbFile extended pipeline support #4783 (#4788)
* Update Get-DbaDbFile.ps1

Adding $SmoDatabase pipeline parameter to support pipelining in an SMO Database object array.  If SmoDatabase is provided, it takes precidence over SqlInstance and Database; they're overwritten with the contents of SmoDatabase.  Added Region block around the queries so I could fold them in VSCode.

* Update Get-DbaDbFile.ps1

Adding Parameter sets to help with the intellisense when writing the function to see that if you specify SmoDatabase, you do not specify SqlInstance or Database

* Adding Parameter information to the function definition, and adding it to the test

Adding new SmoDatabase parameter to the tests, and filling out the .PARAMETER and .EXAMPLE information for the new parameter and functionality.

* Updating test to dynamically handle known param count

removing $paramCount and driving count from $knownParameters.Count

* Replacing ParameterSets with if logic

Using If logic instead of ParameterSets to prevent unexpected errors for users of the function related to ParameterSet violations.

* Formatting updates

Ran DbatoolsFormatter against file, formatting

* updates