Issue 4364 - Query Store Options for 2017 (#5926)
* Issue 4364 - Query Store Options for 2017

Adding the max plan per queries and wait stats capture mode.  As noted code is in here for 2019 features as well but commented out, I'll come back and fix that when it is RTM.  But after many hours of fussing with git I think this will commit.  Yay!

* Issue 4364 - Query Store Options for 2017

Fix formattting

* Fix comments and test

* Invoke-DbatoolsFormatter for functions.

* add 2019 features, do a lil cleanup

* add params