Update Test-DbaLastBackup to specify device type of disk
Modified the calls to Get-DbaBackupHistory to limit the backup results to only that of the disk device type. Third party tools (VMWare Snapshots, Veeam, etc) often use the VSS service while backup up the host system. This creates an entry in the backup tables as if they were a normal backup, but of the virtual device type. Since these backups don't exist in the sense of being usable as a database backup, this change will exclude these and only include the latest full backup set. I do not know of any other device types aside from disk that would be usable in this process, so limited it to just disk. If anyone does know of or need others included, it should be a small change to rope those in too, or make it a paramater.