Added optional User parameter to Set-DbaPrivilege
* In order to prepare for changing a service account, permissions should
be able to be set for the new account ahead of the change in order to
minimize service restarts. Adding an optional parameter for the user to
set allows for this possibility. If provided, it skips the remote
service enumeration and attempts to set the requested permissions for
the passed in user.

* Fixed what appeared to be a bug where... none of the permissions
appeared to be set.