Set-DbaTempDbConfig - Fix the resize of the log file by default #6202
Function Set-DbaTempDbConfig sets the tempdb log file size as 25%
of total data files, even though the parameter LogFileSize is not
specified. With this change the size of the log file is only changed
when that parameter is specified. In any case, in the output of the
command, the final log file size (modified or not) is displayed
correctly