Set-DbaTempDbConfig - Replace function ExecuteWithResults by function Query
To make the code easier to read it is recommended to use the
function $server.Databases[tempdb].Query() instead of function
$server.Databases[tempdb].ExecuteWithResults(), so in this commit
it has been replaced in all places it was used.
Also the output column name FileName has been replaced by PhysicalName and
Name by LogicalName referring to database data and log files.