Fix trying to run Get-DbaDbBackupHistory with empty AvailabilityGroupListener
The recursive call of Get-DbaDbBackupHistory when trying to get
history from replicas has done this via the listener of that
availability group without testing for existance of a listener.

I added a test for existance of a listener to skip the recursive
call if non is found.

In my opinion this is the smallest possible change to solve
the issue #6607.