Find-DbaOrphanedFile: Match local and UNC paths for the same file.
Update logic to work with all paths in UNC format so that local and
UNC references to the same file will match successfully. Filename and
RemoteFilename values in output are unaffected. This commit includes
a new internal function, Split-AdminUnc, which splits a UNC file
path into server name and local path; i.e. the inverse of the
Join-AdminUnc function.