Fix misnamed variable in Backup-DbaDatabase?
If you try to perform a backup on multiple databases with the `-ReplaceInName` and the `-FilePath` flag, you get hit with the error "1 BackupFile specified, but more than 1 database." 

According to the examples, this should be possible: `Backup-DbaDatabase -SqlInstance Server1\Prod -Path \\filestore\backups\servername\instancename\dbname\backuptype -FilePath dbname-backuptype-timestamp.trn -Type Log -ReplaceInName`

I think this is because there's a misnamed variable that doesn't exist anywhere else. When I set `$ReplaceInFile = $true` it works, and when it's false, the command fails and raises the error message. 

I'm pretty sure that `$ReplaceInFile` is meant to be `$ReplaceInName`, but clarification otherwise would also be appreciated.