Update Install-DbaMaintenanceSolution.ps1
Added Queue and QueueDatabase to the list of required scripts for anyone who wants to use the DatabasesInParallel options of the jobs so they don't have to go and find them separately.

QueueDatabase creates an FK to Queue so had to add a Sort-Order on the for-each to ensure the right order of deployment.

Added cleanup task for the two tables for $ReplaceExisting

Changed existence check for the Queue & QueueDatabase to Stop should they exist and not have $ReplaceExisting enabled