Removed secondary as default MonitorServer if Force switch is set
Setting the secondary server as monitor server will break transaction log
shipping if the secondary is not prepared to be set up as a monitor server.

The default should be not to use the secondary as a monitoring server.

The change will skip all dependent checks if no monitoring server is provided.