Temptables can be used on NonPooledConnections
local temptables require the same conneciton to be used when creating
and querying them. Global temptables just needs the connection that
created it to stay open. Prior to this fix all nonPooledConnecions
were disconnected after querying with Invoke-DbaQuery. Also prior to
this fix, Write-DbaDbTableData would create a new connection even if
an open one was given. Now if an open connection is given it will be
reused, and not disconnected.