Fixed connection db switching to tempdb
In Write-DbaDbTableData, there was an issue with creating a temptable
that would set the connection's db to tempdb for the remainder of its
lifetime. This fix checks if a temptable was created, and if so, will
now swap back to the db in use at the start of the command.